Refactoring data reducer
- move business logic to matDispatchToProps
